+++ docker/system/cleanup.sh
@@ -1,0 +1,13 @@
+#!/bin/sh
+
+# include err(), stat() and other functions
+. ./lib.sh
+
+# Kill all controller containers (optionally do `make clean`)
+echo "You may want run make clean as well"
+
+kill1 clixon/clixon-system
+
+
+
+